Split up main module
